# Speed up cupy environment duplicate detection