@kmaehashi 2023-12-07
Hi @hauntsaninja, thanks for the PR! This looks great, I'll test & merge after the release planned today.